% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdds.R
\name{gdds}
\alias{gdds}
\title{Calculation for Growing Degree Days (GDDs, ºC d)}
\usage{
gdds(
  data,
  date_col,
  temp_col,
  group_col = NULL,
  date = c(start_date, end_date),
  BT = 0
)
}
\arguments{
\item{data}{A data frame containing the input data.}

\item{date_col}{The name of the column in the data frame containing date values.}

\item{temp_col}{The name of the column in the data frame containing temperature values.}

\item{group_col}{Optional. The name of the column to group by, if desired.}

\item{date}{A vector of two elements specifying the start and end dates for the calculation (e.g., \code{c("2023-11-23", "2024-01-23")}).}

\item{BT}{Base temperature to subtract from each temperature value. Default is 0. If the temperature is less than Base temperature, it is considered as 0.}
}
\value{
A data frame with the filtered data, including new columns for Days and GDDs. When calculating GDDs, if the temperature is below zero, it is considered as 0.
}
\description{
This function calculates the cumulative growing degree days (GDDs, ºC d) and the sequential days for a specified date range.
}
\examples{
# Example usage with start_date and end_date
# Calculate cumulative temp from 2023-11-23 to 2024-01-23 when the base temperature is 0
cumulative_temp = gdds(df, "date", "temp", date= c("2023-11-23", "2024-01-23"), BT= 0)

# Calculate cumulative temp from 2023-11-23 to 2024-01-23 per group when the base temperature is 0
cumulative_temp_by_group = gdds(df, "date", "temp", "group", date= c("2023-11-23", "2024-01-23"), BT= 0)

}
